fix: create folders needed (#50)
* fix: create folders needed

* check to remove folder

* chore: udpate version

+++ electron/updater.ts
@@ -104,7 +104,6 @@ const reportFatalError = async (sender: WebContents, message: string) => {
 }
 
 const reportCrash = async (sender: WebContents) => {
-
   const path = JSON.stringify({ playerlogpath: getPlayerLogPath() })
   const data = JSON.stringify(`Player log:\n${getPlayerLog()}`)
   const code = `window.Rollbar.error(${data}, ${path})`
@@ -159,7 +158,10 @@ const registerDownloadEvent = (win: BrowserWindow, launcherPaths: LauncherPaths)
   //electronDl();
   ipcMain.on('download', async (event) => {
     const branchPath = launcherPaths.rendererPath + getBranchName()
-    fs.rmdirSync(branchPath, { recursive: true })
+    if (fs.existsSync(branchPath)) {
+      fs.rmSync(branchPath, { recursive: true })
+    }
+    createDirIfNotExists(branchPath)
     const url = launcherPaths.baseUrl + main.config.desktopBranch + launcherPaths.artifactUrl
     console.log('artifactUrl: ', url)
     const res = await electronDl.download(win, url, {
@@ -175,7 +177,9 @@ const registerDownloadEvent = (win: BrowserWindow, launcherPaths: LauncherPaths)
       onCompleted: (file) => {
         console.log('onCompleted:', file)
         unzip(file.path, branchPath, () => {
-          fs.rmSync(file.path)
+          if (fs.existsSync(file.path)) {
+            fs.rmSync(file.path)
+          }
 
           const versionData = {
             version: remoteVersion
@@ -193,8 +197,16 @@ const registerDownloadEvent = (win: BrowserWindow, launcherPaths: LauncherPaths)
   })
 }
 
+const createDirIfNotExists = (path: string) => {
+  if (!fs.existsSync(path)) {
+    fs.mkdirSync(path)
+  }
+}
+
 export const registerUpdaterEvents = (win: BrowserWindow, launcherPaths: LauncherPaths) => {
   try {
+    createDirIfNotExists(launcherPaths.rendererPath)
+
     // Get version
     registerVersionEvent(launcherPaths)
 
@@ -206,7 +218,9 @@ export const registerUpdaterEvents = (win: BrowserWindow, launcherPaths: Launche
 
     // Register clear cache
     ipcMain.on('clearCache', async (event) => {
-      fs.rmdirSync(launcherPaths.rendererPath, { recursive: true })
+      if (fs.existsSync(launcherPaths.rendererPath)) {
+        fs.rmSync(launcherPaths.rendererPath, { recursive: true })
+      }
     })
   } catch (e) {
     console.error('registerUpdaterEvents error: ', e)
